Gaaaaaaaanz viele Änderungen

+++ Bibo Verwaltung/Buch.cs
@@ -79,7 +79,7 @@ namespace Bibo_Verwaltung
         public Buch()
         {
             FillObjectBuch();
-            FillObjectLeihen();
+            FillObject_All();
         }
         //public Buch()
         //{
@@ -89,18 +89,18 @@ namespace Bibo_Verwaltung
         public Buch(string isbn)
         {
             this.isbn = isbn;
-            Load();
+            Load_Buch();
             //FillObject();
-            FillObjectLeihen();
+            FillObject_All();
         }
         #endregion
 
         #region Buch per ISBN laden
-        private void Load()
+        private void Load_Buch()
         {
             //SQL-Verbindung pruefen
             if (con.ConnectError()) return;
-            string RawCommand = "SELECT *, isnull(buch_erscheinungsdatum, '01.01.1990') as 'verified_erscheinungsdatum' FROM t_s_buecher left join t_s_genre on buch_genre_id = ger_id left join t_s_autor on buch_autor_id = au_id left join t_s_verlag on buch_verlag_id = ver_id left join t_s_sprache on buch_sprache_id = sprach_id WHERE buch_isbn = @0";
+            string RawCommand = "SELECT *, isnull(buch_erscheinungsdatum, '01.01.1990') as 'verified_erscheinungsdatum' FROM [dbo].[t_s_buecher] left join [dbo].[t_s_genre] on buch_genre_id = ger_id left join [dbo].[t_s_autor] on buch_autor_id = au_id left join [dbo].[t_s_verlag] on buch_verlag_id = ver_id left join [dbo].[t_s_sprache] on buch_sprache_id = sprach_id WHERE buch_isbn = @0";
             SqlDataReader dr = con.ExcecuteCommand(RawCommand, isbn);
             // Einlesen der Datenzeilen 
             while (dr.Read())
@@ -132,11 +132,12 @@ namespace Bibo_Verwaltung
         }
         #endregion
 
-        #region Buch in Datenbank speichern
-        public void Save()
+        #region Buch-Eigenschaften aendern
+        public void Update_Buch()
         {
-            string RawCommand = "UPDATE [dbo].[t_s_buecher] set buch_titel = @titel , buch_autor_id = @autor, buch_genre_id = @genre, buch_sprache_id = @sprache, buch_verlag_id = @verlag, buch_auflage = @auflage, buch_erscheinungsdatum = @er_datum, buch_neupreis = @neupreis WHERE buch_isbn = @isbn";
+            //SQL-Verbindung pruefen
             if (con.ConnectError()) return;
+            string RawCommand = "UPDATE [dbo].[t_s_buecher] set buch_titel = @titel , buch_autor_id = @autor, buch_genre_id = @genre, buch_sprache_id = @sprache, buch_verlag_id = @verlag, buch_auflage = @auflage, buch_erscheinungsdatum = @er_datum, buch_neupreis = @neupreis WHERE buch_isbn = @isbn";
             SqlCommand cmd = new SqlCommand(RawCommand, con.Con);
 
             cmd.Parameters.AddWithValue("@titel", Titel);
@@ -148,43 +149,41 @@ namespace Bibo_Verwaltung
             cmd.Parameters.AddWithValue("@er_datum", Er_datum);
             cmd.Parameters.AddWithValue("@neupreis", neupreis);
             cmd.Parameters.AddWithValue("@isbn", isbn);
- 
+
             cmd.ExecuteNonQuery();
             con.Close();
         }
         #endregion
 
-        #region Add
-        public void Add()
+        #region Buch hinzufuegen
+        public void Add_Buch()
         {
-            {
-                string RawCommand = "INSERT INTO [dbo].[t_s_buecher] (buch_isbn, buch_titel, buch_genre_id, buch_autor_id, buch_verlag_id, "
-                    + "buch_erscheinungsdatum, buch_sprache_id, buch_auflage, buch_neupreis) VALUES "
-                    + "(@isbn, @titel, @genreid, @autorid, @verlagid, @erscheinungsdatum, @sprachid, @auflage, @neupreis)";
-                con.ConnectError();
-                SqlCommand cmd = new SqlCommand(RawCommand, con.Con);
+            //SQL-Verbindung pruefen
+            if (con.ConnectError()) return;
+            string RawCommand = "INSERT INTO [dbo].[t_s_buecher] (buch_isbn, buch_titel, buch_genre_id, buch_autor_id, buch_verlag_id, buch_erscheinungsdatum, buch_sprache_id, buch_auflage, buch_neupreis) VALUES (@isbn, @titel, @genreid, @autorid, @verlagid, @erscheinungsdatum, @sprachid, @auflage, @neupreis)";
+            SqlCommand cmd = new SqlCommand(RawCommand, con.Con);
 
-                cmd.Parameters.AddWithValue("@isbn", ISBN);
-                cmd.Parameters.AddWithValue("@titel", Titel);
-                cmd.Parameters.AddWithValue("@genreid", Genre.GenreID);
-                cmd.Parameters.AddWithValue("@autorid", Autor.AutorID);
-                cmd.Parameters.AddWithValue("@verlagid", Verlag.VerlagID);
-                cmd.Parameters.AddWithValue("@erscheinungsdatum", Er_datum);
-                cmd.Parameters.AddWithValue("@sprachid", Sprache.SpracheID);
-                cmd.Parameters.AddWithValue("@auflage", Auflage);
-                cmd.Parameters.AddWithValue("@neupreis", Neupreis);
- 
-                cmd.ExecuteNonQuery();
-                con.Close();
-            }
+            cmd.Parameters.AddWithValue("@isbn", ISBN);
+            cmd.Parameters.AddWithValue("@titel", Titel);
+            cmd.Parameters.AddWithValue("@genreid", Genre.GenreID);
+            cmd.Parameters.AddWithValue("@autorid", Autor.AutorID);
+            cmd.Parameters.AddWithValue("@verlagid", Verlag.VerlagID);
+            cmd.Parameters.AddWithValue("@erscheinungsdatum", Er_datum);
+            cmd.Parameters.AddWithValue("@sprachid", Sprache.SpracheID);
+            cmd.Parameters.AddWithValue("@auflage", Auflage);
+            cmd.Parameters.AddWithValue("@neupreis", Neupreis);
+
+            cmd.ExecuteNonQuery();
+            con.Close();
         }
         #endregion
 
-        #region Buch aus Datenbank loeschen (per ISBN)
-        public void Delete()
+        #region Buch loeschen (per ISBN)
+        public void Delete_Buch()
         {
+            //SQL-Verbindung pruefen
+            if (con.ConnectError()) return;
             string RawCommand = "DELETE FROM [dbo].[t_s_buecher] WHERE buch_isbn = @isbn";
-            con.ConnectError();
             SqlCommand cmd = new SqlCommand(RawCommand, con.Con);
             cmd.Parameters.AddWithValue("@isbn", ISBN);
             cmd.ExecuteNonQuery();
@@ -197,7 +196,7 @@ namespace Bibo_Verwaltung
         DataSet ds = new DataSet();
         DataTable dt = new DataTable();
         SqlCommandBuilder comb = new SqlCommandBuilder();
-        private void FillObjectLeihen()
+        private void FillObject_All()
         {
             SQL_Verbindung con = new SQL_Verbindung();
             if (con.ConnectError()) return;
@@ -215,9 +214,7 @@ namespace Bibo_Verwaltung
             adapter.Fill(dt);
             con.Close();
         }
-        #endregion
-
-        #region Fill ObjectBuch
+       
         SqlDataAdapter adapter2 = new SqlDataAdapter();
         DataSet ds2 = new DataSet();
         DataTable dt2 = new DataTable();
@@ -227,55 +224,87 @@ namespace Bibo_Verwaltung
         {
             if (con2.ConnectError()) return;
             //string RawCommand = "SELECT buch_isbn, buch_titel, buch_erscheinungsdatum, buch_auflage, buch_neupreis, ger_name, au_autor, ver_name, sprach_name FROM t_s_buecher left join t_s_genre on buch_genre_id = ger_id left join t_s_autor buch_autor_id = au_id left join t_s_verlag buch_verlag_id = ver_id left join t_s_sprache buch_sprache = sprach_id";
-            string RawCommand = "SELECT buch_isbn as 'ISBN',"
-                + "buch_titel as 'Titel',"
-                + "ger_name as 'Genre',"
-                + "au_autor as 'Autor',"
-                + "ver_name as 'Verlag',"
-                + "buch_erscheinungsdatum as 'Erscheinungsdatum',"
-                + "sprach_name as 'Sprache',"
-                + "buch_auflage as 'Auflage',"
-                + "buch_neupreis as 'Neupreis' from t_s_buecher "
+            string RawCommand = "SELECT * from t_s_buecher "
                 + "left join t_s_genre on buch_genre_id = ger_id "
                 + "left join t_s_autor on buch_autor_id = au_id "
                 + "left join t_s_verlag on buch_verlag_id = ver_id "
                 + "left join t_s_sprache on buch_sprache_id = sprach_id";
+            //string RawCommand = "SELECT buch_isbn as 'ISBN',"
+            //    + "buch_titel as 'Titel',"
+            //    + "ger_name as 'Genre',"
+            //    + "au_autor as 'Autor',"
+            //    + "ver_name as 'Verlag',"
+            //    + "buch_erscheinungsdatum as 'Erscheinungsdatum',"
+            //    + "sprach_name as 'Sprache',"
+            //    + "buch_auflage as 'Auflage',"
+            //    + "buch_neupreis as 'Neupreis' from t_s_buecher "
+            //    + "left join t_s_genre on buch_genre_id = ger_id "
+            //    + "left join t_s_autor on buch_autor_id = au_id "
+            //    + "left join t_s_verlag on buch_verlag_id = ver_id "
+            //    + "left join t_s_sprache on buch_sprache_id = sprach_id";
             adapter = new SqlDataAdapter(RawCommand, con2.Con);
             adapter.Fill(ds2);
             adapter.Fill(dt2);
             con2.Close();
         }
         #endregion
+        //TODO
 
-        #region ClearDS()
+        #region DataSet zuruecksetzen
         public void ClearDSBuch()
         {
             ds2.Tables[0].Rows.Clear();
         }
         #endregion
+        //TODO
 
-        #region FillGrid
-        public void FillGridBuch(ref DataGridView grid, object value = null)
+        #region DataGridView fuellen
+        public void FillGrid_Buch(ref DataGridView grid, object value = null)
         {
             grid.DataSource = ds2.Tables[0];
+
+            grid.Columns["buch_isbn"].HeaderText = "ISBN";
+            grid.Columns["buch_titel"].HeaderText = "Titel:";
+            grid.Columns["ger_name"].HeaderText = "Genre:";
+            grid.Columns["au_autor"].HeaderText = "Autor:";
+            grid.Columns["ver_name"].HeaderText = "Verlag:";
+            grid.Columns["buch_erscheinungsdatum"].HeaderText = "Erscheinungsdatum:";
+            grid.Columns["sprach_name"].HeaderText = "Sprache:";
+            grid.Columns["buch_auflage"].HeaderText = "Auflage:";
+            grid.Columns["buch_neupreis"].HeaderText = "Neupreis:";
+
+            grid.Columns["buch_titel"].HeaderText = "Titel:";
+            grid.Columns["buch_titel"].HeaderText = "Titel:";
+            grid.Columns["buch_titel"].HeaderText = "Titel:";
+            grid.Columns["buch_titel"].HeaderText = "Titel:";
+
+            grid.Columns["buch_titel"].HeaderText = "Titel:";
         }
 
-        public void FillGrid(ref DataGridView grid, object value = null)
+        public void FillGrid_Load_All(ref DataGridView grid, object value = null)
         {
             grid.DataSource = ds.Tables[0];
-            //grid.Columns[2].Visible = false;
-            //grid.Columns[4].Visible = false;
-            //grid.Columns[6].Visible = false;
-            //grid.Columns[7].Visible = false;
-            //grid.Columns[8].Visible = false;
-            //grid.Columns[10].Visible = false;
-            //grid.Columns[13].Visible = false;
-            //grid.Columns[15].Visible = false;
-            //grid.Columns[17].Visible = false;
-            //grid.Columns[19].Visible = false;
-            //grid.Columns[21].Visible = false;
-            //grid.Columns[25].Visible = false;
-            //grid.Columns[28].Visible = false;
+            grid.Columns[2].Visible = false;
+            grid.Columns[4].Visible = false;
+            grid.Columns[6].Visible = false;
+            grid.Columns[7].Visible = false;
+            grid.Columns[8].Visible = false;
+            grid.Columns[10].Visible = false;
+            grid.Columns[13].Visible = false;
+            grid.Columns[15].Visible = false;
+            grid.Columns[17].Visible = false;
+            grid.Columns[19].Visible = false;
+            grid.Columns[21].Visible = false;
+            grid.Columns[24].Visible = false;
+            grid.Columns[27].Visible = false;
+            grid.Columns[29].Visible = false;
+            grid.Columns[32].Visible = false;
+            grid.Columns[33].Visible = false;
+            grid.Columns[34].Visible = false;
+            grid.Columns[35].Visible = false;
+            grid.Columns[37].Visible = false;
+            grid.Columns[38].Visible = false;
+            grid.Columns[39].Visible = false;
 
             grid.Columns["bu_id"].HeaderText = "Buch-ID:";
             grid.Columns["bu_isbn"].HeaderText = "Buch-ISBN:";
@@ -292,7 +321,10 @@ namespace Bibo_Verwaltung
             grid.Columns["aus_leihnummer"].HeaderText = "Ausleihnummer:";
             grid.Columns["aus_leihdatum"].HeaderText = "ausgeliehen am:";
             grid.Columns["aus_rückgabedatum"].HeaderText = "Rückgabe am:";
-
+            grid.Columns["aus_barcode"].HeaderText = "Barcode:";
+            grid.Columns["kunde_vorname"].HeaderText = "Vorname:";
+            grid.Columns["kunde_nachname"].HeaderText = "Nachname:";
+            grid.Columns["kunde_vertrauenswürdigkeit"].HeaderText = "Vertrauenswürdigkeit:";
         }
         #endregion
 
@@ -315,14 +347,14 @@ namespace Bibo_Verwaltung
                     ergebnis = false;
                 }
             }
-            dr.Close(); 
+            dr.Close();
             con.Close();
             return ergebnis;
         }
 
-        public void Ausleihen(int buch_id, string aus_datum, string rueck_datum, int kunde)
+        public void Ausleihen(int buch_id, string aus_datum, string rueck_datum, int kunde, string barcode)
         {
-            string RawCommand = "INSERT into [dbo].[t_bd_ausgeliehen] (aus_buchid, aus_leihdatum, aus_rückgabedatum, aus_kundenid) VALUES(@buchid, @ausdatum, @rückdatum, @kunde)";
+            string RawCommand = "INSERT into [dbo].[t_bd_ausgeliehen] (aus_buchid, aus_leihdatum, aus_rückgabedatum, aus_kundenid, aus_barcode) VALUES(@buchid, @ausdatum, @rückdatum, @kunde, @barcode)";
             if (con.ConnectError()) return;
             SqlCommand cmd = new SqlCommand(RawCommand, con.Con);
 
@@ -330,7 +362,8 @@ namespace Bibo_Verwaltung
             cmd.Parameters.AddWithValue("@ausdatum", aus_datum);
             cmd.Parameters.AddWithValue("@rückdatum", rueck_datum);
             cmd.Parameters.AddWithValue("@kunde", kunde);
- 
+            cmd.Parameters.AddWithValue("@barcode", barcode);
+
             cmd.ExecuteNonQuery();
             con.Close();
         }
@@ -340,7 +373,7 @@ namespace Bibo_Verwaltung
             if (con.ConnectError()) return;
             string RawCommand = "DELETE FROM t_bd_ausgeliehen WHERE aus_buchid = @0";
             SqlCommand cmd = new SqlCommand(RawCommand, con.Con);
-            cmd.Parameters.AddWithValue("@0", buch_id);   
+            cmd.Parameters.AddWithValue("@0", buch_id);
             cmd.ExecuteNonQuery();
             con.Close();
         }
